==> notebooks/01-data-wrangling/01-mtgjson-data/00-mtgjson-data-intro.ipynb <==
 # Summary

 An introduction to the [MTGJSON](https://mtgjson.com/) card data, including reviewing the data files and column features.

## Introduction

I will use the data from [MTGJSON](https://mtgjson.com/downloads/all-files/).  The `AllPrintings` card data comes in various formats, such as json, sql, csv, and parquet.

I will use the [parquet format](https://parquet.apache.org/), since that is the most performant format for data analysis.  It has high compression, fast load times, and can query directly on disk.  This minimizes both disk space and memory usage.

I will also use the `AllPrices` data for economic analysis.  This is only available in the json format.

## Downloads

See notebook [10-get-data.ipynb](10-get-data.ipynb) to fetch the data.


# Review AllPrintings Tables

We have 18 parquet files associated with the card data, let't take a quick tour.  

# Setup Notebook
import os
if os.path.basename(os.getcwd()) != "mtg-modeling":
    get_ipython().run_line_magic("run", '-i "../../scripts/notebook_header.py"')  # type: ignore

# list of files
save_path = Path("data/raw/mtgjson")
file_base = "AllPrintingsParquetFiles"  # extension added later
final_path = save_path / file_base

path = final_path
files = os.listdir(path)
files.sort()
files

## Card Files:
- `cards.parquet`: The primary file that contains card data, such as card name, mana cost, type, and text.
- `tokens.parquet`: Same for tokens.
- `cardForeignData.parquet`: Foreign language translations of cards.
- `cardLegalities.parquet`: Legality of cards for various play formats.
- `cardPrices.parquet`: Latest prices for cards on various platforms, including retail and buylist prices.
- `cardPurchaseUrls.parquet`: URLs to various retail platforms.
- `cardRulings.parquet`: The rulings for cards.

## Set Files:
- `sets.parquet`: Data on various released sets, such as set code (10E, OTJ...), set size, and release date.
- `setTranslations.parquet`: Translations for set names in various languages.

## Identifier Files:
- `cardIdentifiers.parquet`: Identifiers for various MTG data platforms (TCG Collector, Scryfall, Cardmarket...).
- `tokenIdentifiers.parquet`: Same for tokens.

## Set Booster Files:
- `setBoosterContents.parquet`: For booster packs, different mixes of sheet composition (1 theList + 13 others versus 0 theList + 14 others).
- `setBoosterContentWeights.parquet`: The weight of each booster mix (1 in 10 boosters has theList).
- `setBoosterSheets.parquet`: Card sheet information.
- `setBoosterSheetCards.parquet`: Card composition of each sheet, including counts.

## Meta File:
- `meta.parquet`: Version and date for current MTGJSON build.


# Unique Identifiers

Most of the files have a `uuid`. This is the universally unique identifier (UUID v5) for each card printing.  It is the primary key for the `cards.parquet` file and will be used to join data across tables.

## MTGJSON
 - `uuid`:
   - Reprinted card editions: Unique id
   - [Double-faced cards](https://mtg.fandom.com/wiki/Double-faced_card) (DBC): Each face has a unique `uuid`.
   - Foreign languages: Same Id. 

## WOTC Gatherer
 - `multiverseId`: The WOTC card identifier used their [Gatherer](https://gatherer.wizards.com) card database.  
    - Reprinted card editions: Unique id
    - Double-faced cards: Same id
    - Foreign languages: Different id

## Scryfall
 - `scryfallId`: The [Scryfall](https://scryfall.com/) uuid.  It has different rules than the MTGJSON uuid, such as faces of DFCs are not unique.
    - Reprinted card editions: Unique id
    - Double-faced cards: Same id.  See `scryfallCardBackId`.
    - Foreign languages: Different id
