==> notebooks/01-data-wrangling/01-mtgjson-data/10-get-data.ipynb <==
# Summary

The code to acquire and pre-process the card data to use for future analysis.

# Grab MTGJSON Card Data

Here I will download and clean the data for MTG cards.

First we will download the data from [MTGJSON](https://mtgjson.com/downloads/all-files/).  The `AllPrintings` card data comes in various formats, such as json, sql, csv, and parquet.

I will use the [parquet format](https://parquet.apache.org/), since that is the most performant format for data analysis.  It has high compression, fast load times, and can query directly on disk.  This minimizes both disk space and memory usage.

Data URLs:
- __Linux__: https://mtgjson.com/api/v5/AllPrintingsParquetFiles.tar.gz
- __Windows__: https://mtgjson.com/api/v5/AllPrintingsParquetFiles.zip

The following code downloads and decompresses the data.

# Setup Notebook
import os
if os.path.basename(os.getcwd()) != 'mtg-modeling':
    get_ipython().run_line_magic("run", '-i "../../scripts/notebook_header.py"') # type: ignore

from src.data.mtgjson_fetcher import MtgJsonFetcher


card_fetcher = MtgJsonFetcher(dataset='AllPrintingsParquetFiles', save_root='data/raw/mtgjson')
card_fetcher.fetch()

# Grab MTGJSON All Price Data

Next we will download the data from [MTGJSON](https://mtgjson.com/downloads/all-files/).  The `AllPrices` card data only comes in json format.  We will have to convert this to parquet for ease of use in future analysis.  

Note this data only covers the previous 90 days.

Since the file is very large, I will use [polars](https://docs.pola.rs) instead of [pandas](https://pandas.pydata.org/docs/)

Data URLs:
- __Linux__: https://mtgjson.com/api/v5/AllPrices.json.gz
- __Windows__: https://mtgjson.com/api/v5/AllPrices.json.zip

The following code downloads and decompresses the data.

price_fetcher = MtgJsonFetcher(dataset='AllPrices.json', save_root='data/raw/mtgjson')
price_fetcher.fetch()
