==> notebooks/01-data-wrangling/01-mtgjson-data/20-card-data-wrangle.ipynb <==
# Summary

Wrangle the all card data.  Joins set data as well.  Then selects Standard Legal cards and other 
interesting data sets.

# Setup Notebook
import os
if os.path.basename(os.getcwd()) != "mtg-modeling":
    get_ipython().run_line_magic("run", '-i "../../scripts/notebook_header.py"')  # type: ignore

import pandas as pd
import polars as pl

SET_CODE = "BLB"

paths = {
    "raw": Path("data/raw/mtgjson/AllPrintingsParquetFiles"),
    "interim": Path("data/interim/mtgjson/AllPrintings"),
    "processed": Path("data/processed/mtgjson/AllPrintings"),
}

paths['raw'].exists()
os.makedirs(paths["interim"], exist_ok=True)
os.makedirs(paths["processed"], exist_ok=True)

cards = pl.scan_parquet(paths["raw"] / "cards.parquet")
legalities = pl.scan_parquet(paths["raw"] / "cardLegalities.parquet")
purchase_urls = pl.scan_parquet(paths["raw"] / "cardPurchaseUrls.parquet")
sets = pl.scan_parquet(paths["raw"] / "sets.parquet")

cards.collect().columns

results = (
    cards
    .join(sets, left_on='setCode', right_on='code', how='left')
    .join(legalities, on='uuid', how='left')
    .join(purchase_urls, on='uuid', how='left')
)

wide_cards = results.collect()
wide_cards.write_parquet(paths["interim"] / "wide_cards.parquet")
print(wide_cards.shape)

## Data for Standard Legal Cards

Used for analysis of the Standard format.  

See the [MTG Wiki Standard/Timeline](https://mtg.fandom.com/wiki/Standard/Timeline) page for validate correct card composition.

standard_results = (
    results
    .filter(pl.col('standard') == 'Legal')
    .filter(pl.col('borderColor') == 'black')
    .filter(pl.col('isPromo').is_null())
    # .filter(pl.col('isReprint').is_null())
    .filter(pl.col('promoTypes').is_null())
    .sort(pl.col('name'), descending=False)
    .sort(pl.col('releaseDate'), descending=False)
    .sort(pl.col('power'), descending=False)
)

standard_cards = standard_results.collect()
standard_cards.write_parquet(paths["processed"] / "standard_cards.parquet")
print(standard_cards.shape)

## Outlaws of Thunder Junction Standard Legal Set

Per the [wiki](https://mtg.fandom.com/wiki/Standard/Timeline), there should be 271 cards in the OTJ set.

otj_results = (
    results
    .filter(pl.col('standard') == 'Legal')
    .filter(pl.col('borderColor') == 'black')
    .filter(pl.col('isPromo').is_null())
    # .filter(pl.col('isReprint').is_null())
    .filter(pl.col('promoTypes').is_null())
    .filter(pl.col('setCode') == SET_CODE)
    .sort(pl.col('number'))
)

otj_cards = otj_results.collect()
otj_cards.write_parquet(paths["processed"] / f"{SET_CODE}_std_cards.parquet")
otj_cards_ids = (
    otj_cards.select([pl.col("name"), pl.col("number").str.zfill(3)])
    .sort(pl.col("number"))
    .to_pandas()
)

otj_core_results = (
    otj_results
    .select([
        pl.col('name'),
        pl.col('setCode'),
        pl.col('releaseDate'),
        pl.col('number').str.zfill(3),
        pl.col('layout'),
        pl.col('availability'),
        pl.col('power'),
        pl.col('toughness'),
        pl.col('colorIdentity'),
        pl.col('colors'),
        pl.col('types'),
        pl.col('subtypes'),
        pl.col('supertypes'),
        pl.col('manaCost'),
        pl.col('manaValue'),
        pl.col('edhrecRank'),
        pl.col('edhrecSaltiness'),
        pl.col('text'),
        pl.col('flavorText'),
    ])
)

otj_core_cards = otj_core_results.collect()
otj_core_cards.write_parquet(paths["processed"] / f"{SET_CODE}_std_thin.parquet")
otj_core_cards.head()
