==> notebooks/01-data-wrangling/01-mtgjson-data/21-booster-data-wrangle.ipynb <==
# Summary

Wrangle the booster set data.

# Introduction

This notebook will wrangle the data needed to understand booster pack composition.

This data is sourced by MTGJSON from [https://github.com/taw/magic-sealed-data](https://github.com/taw/magic-sealed-data).  See the [about](../../00-intro/10-introduction.ipynb) page.


I made a dataset for simulating booster packs.  This will look at for each combination
of booster pack config, sheet name, and card name. The three rates that affect expected pull rates 
are:
- Booster Pack Config: Unique mixes of booster packs, such as some having "The List" cards or 
land foils.
- Sheet Picks: Number of picks from each sheet type in the booster config, such as some packs having 
6 picks from the commons sheets, while others get 5 picks.
- Card Sheet Rate: The rate at which a card uuid occurs on the sheet.
These rates are then used to create an expected pull rate.   I break this into two expected pull 
rate.  One is per uuid, which has different values for alternate cards.  The other is per name, 
which lumps all card alternates with the same gameplay mechanics together.

I took the above dataset and collapse it to the expected pull rates for each card and 
uuid, aggregating the pack configurations and sheet names.  This can be added to other card stats.


# Initial Load and Transforms



# Setup Notebook
import os
if os.path.basename(os.getcwd()) != "mtg-modeling":
    get_ipython().run_line_magic("run", '-i "../../scripts/notebook_header.py"')  # type: ignore

import pandas as pd
import polars as pl

Set paths.  Using the OUTLAWS OF THUNDER JUNCTION set code and play booster type.

SET_CODE = "BLB" # OUTLAWS OF THUNDER JUNCTION
BOOSTER_NAME = "play" # play boosters

paths = {
    "raw": Path("data/raw/mtgjson/AllPrintingsParquetFiles"),
    "interim": Path("data/interim/mtgjson/Boosters"),
    "processed": Path("data/processed/mtgjson/Boosters"),
}

paths['raw'].exists()
os.makedirs(paths["interim"], exist_ok=True)
os.makedirs(paths["processed"], exist_ok=True)

Scan files into lazy dataframes.  No compute until .collect() method is called.

cards = pl.scan_parquet(
    Path(f"data/processed/mtgjson/AllPrintings/{SET_CODE}_std_cards.parquet")
)
boost_content = pl.scan_parquet(paths["raw"] / "setBoosterContents.parquet")
boost_weight = pl.scan_parquet(paths["raw"] / "setBoosterContentWeights.parquet")
boost_sheets = pl.scan_parquet(paths["raw"] / "setBoosterSheets.parquet")
boost_cards = pl.scan_parquet(paths["raw"] / "setBoosterSheetCards.parquet")

Tidy the card data.

df_cards = (
    cards.select(["name", "setCode", "number", "uuid"])
    .filter(pl.col("setCode") == SET_CODE)
    .join(boost_cards, left_on="uuid", right_on="cardUuid", how="inner")
    .with_columns(
        [
            pl.col("number")
            .str.zfill(3)
            .alias("number")
        ]
    )
)
df_cards.collect().head()

Tidy the booster composition, which shows which sheet names and frequencies occur in a give booster.

Display the pivot to compare different booster compositions.

df_bc = boost_content.filter(pl.col("setCode") == SET_CODE).filter(
    pl.col("boosterName") == BOOSTER_NAME
)
df_bc.collect().pivot(values="sheetPicks", on="boosterIndex")

Tidy the booster weight.  Shows the frequency at which each booster config occurs.

df_wgt = (
    boost_weight.filter(pl.col("setCode") == SET_CODE)
    .filter(pl.col("boosterName") == BOOSTER_NAME)
    # .sort(pl.col("sheetName"))
    .with_columns(
        (pl.col("boosterWeight") / pl.col("boosterWeight").sum()).alias("boosterConfigRate"),
    )
)
df_wgt.collect().to_pandas()

Tidy the booster sheets data.  This has some data for each sheet type.

df_bs = boost_sheets.filter(pl.col("setCode") == SET_CODE).filter(
    pl.col("boosterName") == BOOSTER_NAME
)
df_bs.collect().to_pandas()

Tidy the booster sheet card data.  This contains the card uuid frequencies on each sheet.

df_bcards = (
    boost_cards.filter(pl.col("setCode") == SET_CODE)
    .filter(pl.col("boosterName") == BOOSTER_NAME)
    .join(
        cards.select(["uuid", "name", "number", "rarity"]),
        left_on="cardUuid",
        right_on="uuid",
        how="inner",
    )
    .with_columns(pl.col("cardWeight").sum().over("sheetName").alias("totalCardWeight"))
    .with_columns((pl.col("cardWeight") / pl.col("totalCardWeight")).alias("cardSheetRate"))
)

df_bcards.collect().to_pandas().sample(10).sort_values("cardSheetRate")

# Draw Rates per Card-Sheet-Booster Combos

Use this for simulating booster pack draws, given rates for booster configs, sheets, and cards.

df = (
    df_bc.join(df_wgt, on=["boosterIndex", "boosterName", "setCode"], how="inner")
    .join(df_bs, on=["sheetName", "boosterName", "setCode"], how="inner")
    .join(df_bcards, on=["sheetName", "boosterName", "setCode"], how="inner")
    .with_columns(
        (
            pl.col("cardSheetRate") * pl.col("boosterConfigRate") * pl.col("sheetPicks")
        ).alias("expectedConfigCardPullRate"),
    )
    .with_columns(
        pl.col("expectedConfigCardPullRate")
        .sum()
        .over("cardUuid")
        .alias("expectedCardUuidPullRate")
    )
    .with_columns(
        pl.col("expectedConfigCardPullRate")
        .sum()
        .over("name")
        .alias("expectedCardNamePullRate")
    )
    .sort(pl.col("expectedCardNamePullRate"), descending=False)
)
df.collect().write_parquet(paths["processed"] / f"{SET_CODE}_booster_sheet_card_rates.parquet")
print(df.collect().shape)
df.filter(
    pl.col("expectedCardUuidPullRate") != pl.col("expectedCardNamePullRate")
).collect().to_pandas().sample(5)

# Draw Rates per Card, Aggregated

Use this for per-card draw rates.  Takes above dataset and aggregates the booster configs and sheets types.

df_card_rates = (
    df.group_by(["cardUuid", "boosterName"])
    .first()
    .select(
        [
            "cardUuid",
            "name",
            "number",
            "setCode",
            "boosterName",
            "expectedCardUuidPullRate",
            "expectedCardNamePullRate",
        ]
    )
)
df_card_rates.collect().write_parquet(
    paths["processed"] / f"{SET_CODE}_card_pull_rates.parquet"
)
df_card_rates.collect().sort(
    "expectedCardUuidPullRate", pl.col("number").str.zfill(3)
).to_pandas().head(5)
