==> notebooks/01-data-wrangling/01-mtgjson-data/22-price-data-wrangle.ipynb <==
# Summary

The code to acquire and pre-process the all price data to use for future analysis.

# Setup Notebook
import os
if os.path.basename(os.getcwd()) != "mtg-modeling":
    get_ipython().run_line_magic("run", '-i "../../scripts/notebook_header.py"')  # type: ignore

import pandas as pd
import polars as pl
from src.data.mtgjson_wrangler import MtgPricesJsonWrangler

PROCESS_DATA = True

filename = "flat_prices_2024-05-25_2024-08-24.parquet"

paths = {
    "raw_file": Path("data/raw/mtgjson/AllPrices/AllPrices.json"),
    "interim": Path("data/interim/mtgjson/AllPrices"),
    "interim_cards": Path("data/interim/mtgjson/AllPrintings"),
    "processed": Path("data/processed/mtgjson/AllPrices"),
}

if PROCESS_DATA:
    wrangler = MtgPricesJsonWrangler(paths)
    wrangler.raw_json_to_parquet()
    wrangler.unstack_data()
else:
    wrangler = MtgPricesJsonWrangler(paths, filename=filename)

df = wrangler.load_data()
df.shape

df_cards = (
    pl.scan_parquet(paths["interim_cards"] /  "wide_cards.parquet")
    .select(["uuid", "name", "setCode", "number", "rarity", "type", "colors"])
    .collect()
)
df_cards.shape

df_full = df.join(df_cards, on="uuid", how="inner")
print(f"Shape: {df_full.shape}")
print(f"Mem Size in GB: {df_full.estimated_size() / (1024**3):.2f}")
df_full.head(1)

min_date = df_full["date"].min()
max_date = df_full["date"].max()
df_full.write_parquet(paths["interim"] / f"wide_prices_{min_date}_{max_date}.parquet")

df_otj = df_full.filter(
    pl.col("setCode") == "OTJ"
)
df_otj.write_parquet(paths["processed"] / f"prices_OTJ_{min_date}_{max_date}.parquet")
df_otj.shape
